% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-aggregate.R
\name{aggregateCountsNoNeg}
\alias{aggregateCountsNoNeg}
\title{Aggregate non-negative probe counts to target level for feature data}
\usage{
aggregateCountsNoNeg(object, FUN_neg = identity, FUN_nonneg = sum)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object to aggregate}

\item{FUN_neg}{function to use for count aggregation in negative probes}

\item{FUN_nonneg}{function to use for count aggregation in non-negative probes}
}
\value{
a NanoStringGeoMxSet object with targets as features
}
\description{
Aggregate non-negative probe counts to target level for feature data
}
\examples{
datadir <- system.file("extdata", "DSP_NGS_Example_Data",
                       package="GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))
targetGeoMxSet <- aggregateCountsNoNeg(demoData, FUN_neg = identity, FUN_nonneg = sum)

}
