% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNanoStringGeoMxSet.R
\name{analyteSubset}
\alias{analyteSubset}
\title{Subset GeomxSet Object by AnalyteType}
\usage{
analyteSubset(object, analyte)
}
\arguments{
\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{analyte}{analyte for returned object}
}
\value{
list containing multiple GeomxSet objects named by AnalyteType
}
\description{
Subset GeomxSet Object by AnalyteType
}
\examples{
testData <- readRDS(file= system.file("extdata","DSP_Proteogenomics_Example_Data", 
"proteinData.rds", package = "GeomxTools"))

proteinData <- analyteSubset(object = aggTestData, analyte = "protein")
RNAData <- analyteSubset(object = aggTestData, analyte = "RNA")

}
