% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-normalize.R
\name{checkQCFlags,NanoStringGeoMxSet-method}
\alias{checkQCFlags,NanoStringGeoMxSet-method}
\title{checkQCFlags}
\usage{
\S4method{checkQCFlags}{NanoStringGeoMxSet}(object, removeLowLocalOutliers = FALSE, ...)
}
\arguments{
\item{removeLowLocalOutliers}{logical, if TRUE it sets outlier counts to zero,  default is FALSE,}

\item{NanoStringGeoMxSet}{}
}
\value{
NanoStringGeoMxSet
}
\description{
checkQCFlags
}
\examples{
QCobject <- checkQCFlags(demoData)
}
