% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNanoStringGeoMxSet.R
\name{compareToConfig}
\alias{compareToConfig}
\title{Compare given PKC probes to probes in config file}
\usage{
compareToConfig(config, pkcProbes, pkcHeader)
}
\arguments{
\item{config}{file path to config file}

\item{pkcProbes}{probe information from readPKCFile}

\item{pkcHeader}{pkc metadata from readPKCFile}
}
\description{
Check if extra PKCs are given based on probes in config file
}
