% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-normalize.R
\name{compute_normalization_factors}
\alias{compute_normalization_factors}
\title{protein normalization}
\usage{
compute_normalization_factors(
  object,
  igg.names = NULL,
  hk.names = NULL,
  area = NULL,
  nuclei = NULL
)
}
\description{
protein normalization
}
