% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-autoplot.R
\name{concordancePlot}
\alias{concordancePlot}
\title{Produce concordance pairs plots}
\usage{
concordancePlot(dat, plotFactor)
}
\arguments{
\item{dat}{dataframe of values to be compared. Could be e.g. IgG probe counts,
or different normalization factors. Each variable/factor is in a column.}

\item{plotFactor}{segment factor to color the plot by}
}
\description{
Draws a pairs plot of concordance between multiple factors/ probes
Upper panels are the concordance plot.
Lower panels are the standard deviation of the log2-ratios between the targets
}
