% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countsShiftedByOne}
\alias{countsShiftedByOne}
\title{Accessor to check if "exprs" \code{assDataElement} was shifted by one}
\usage{
countsShiftedByOne(object)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}
}
\value{
boolean indicating if counts in default matrix were shifted by one
}
\description{
Accessor to check if "exprs" \code{assDataElement} was shifted by one
}
\examples{
countsShiftedByOne(demoData)

}
