% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hk_names}
\alias{hk_names}
\title{Return the House Keeper positive controls for protein}
\usage{
hk_names(object)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}
}
\value{
names of HKs
}
\description{
Return the House Keeper positive controls for protein
}
