% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hkNames}
\alias{hkNames}
\title{Return the House Keeper positive controls for protein}
\usage{
hkNames(object)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}
}
\value{
names of HKs
}
\description{
Return the House Keeper positive controls for protein
}
