% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igg_names}
\alias{igg_names}
\title{Return the IgG negative controls for protein}
\usage{
igg_names(object)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}
}
\value{
names of IgGs
}
\description{
Return the IgG negative controls for protein
}
