% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iggNames}
\alias{iggNames}
\title{Return the IgG negative controls for protein}
\usage{
iggNames(object)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}
}
\value{
names of IgGs
}
\description{
Return the IgG negative controls for protein
}
