% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logtBase}
\alias{logtBase}
\title{Get take the log of a numeric vector}
\usage{
logtBase(x, thresh = 0.5, base = 2)
}
\arguments{
\item{x}{numeric vector}

\item{thresh}{minimum numeric value greater than 0 to have in vector}

\item{base}{numeric value indicating base to log with}
}
\value{
numeric vector with logged values
}
\description{
Get take the log of a numeric vector
}
\examples{
logtBase(c(0, 1, 2, 2), thresh=0.1, base=10)

}
