% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-de.R
\name{mixedModelDE}
\alias{mixedModelDE}
\title{Run a mixed model on GeoMxSet}
\usage{
mixedModelDE(
  object,
  elt = "exprs",
  modelFormula = NULL,
  groupVar = "group",
  nCores = 1,
  multiCore = TRUE,
  pAdjust = "BY",
  pairwise = TRUE
)
}
\arguments{
\item{object}{name of the object class to perform QC on
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{elt}{assayDataElement of the geoMxSet object to run the DE on}

\item{modelFormula}{formula used in DE, if null, the design(object) is used}

\item{groupVar}{= "group",  sample annotation to group the data for comparing means}

\item{nCores}{= 1, number of cores to use, set to 1 if running in serial mode}

\item{multiCore}{= TRUE, set to TRUE to use multiCore, FALSE to run in cluster mode}

\item{pAdjust}{= "BY" method for p-value adjustment}
}
\value{
mixed model output list
}
\description{
Run a mixed model on GeoMxSet
}
\examples{
datadir <- system.file("extdata", "DSP_NGS_Example_Data",
                       package="GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))
target_demoDataSub <- aggregateCounts(demoData)
target_demoDataSub <- target_demoDataSub[1:10, 1:10]
mixedOutmc <- mixedModelDE(target_demoDataSub,
                         elt = "q_norm",
                         modelFormula = ~ pool_rep +  (1 | slide),
                         groupVar = "pool_rep",
                         nCores = 12,
                         multiCore = TRUE,
                         pAdjust = NULL)

}
