% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ngeoMean}
\alias{ngeoMean}
\title{Get the geometric mean of a vector}
\usage{
ngeoMean(x, thresh = 0.5)
}
\arguments{
\item{x}{numeric vector}

\item{thresh}{minimum numeric value greater than 0 to have in vector}
}
\value{
numeric geometric mean of vector
}
\description{
Get the geometric mean of a vector
}
\examples{
ngeoMean(c(0, 1, 2, 2), thresh=0.1)

}
