% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Generate concordance figure of targets based on user provided factors}
\usage{
plot_concordance(targetList, object, plot_factors)
}
\arguments{
\item{targetList}{names of targets to plot concordance, normally IgGs.}

\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{plot_factors}{segment factor to color the plot by}
}
\description{
Upper panels are the concordance plot.
Lower panels are the standard deviation of the log2-ratios between the targets
}
\examples{
testData <- readRDS(file= system.file("extdata","DSP_Proteogenomics_Example_Data", 
"proteinData.rds", package = "GeomxTools"))

proteinData <- analyteSubset(object = aggTestData, analyte = "protein")
RNAData <- analyteSubset(object = aggTestData, analyte = "RNA")

igg.names <- igg_names(proteinData)

plot_concordance(targetList = igg.names, object = proteinData, plot_factors = "Segment_Type)
plot_concordance(targetList = c("C1orf43", "GPI", "OAZ1"), object = RNAData, plot_factors = "Segment_Type)

}
