% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{plot_normFactor_concordance}
\alias{plot_normFactor_concordance}
\title{Add QC flags to feature and protocol data simultaneously}
\usage{
plot_normFactor_concordance(object, plot_factors, normfactors = NULL)
}
\description{
Add QC flags to feature and protocol data simultaneously
}
