% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-autoplot.R
\name{print.sd.log.ratio}
\alias{print.sd.log.ratio}
\title{Prints the standard deviation of log2-ratios}
\usage{
\method{print}{sd.log.ratio}(data, mapping)
}
\arguments{
\item{data}{dataframe of XY values}

\item{mapping}{ggplot mapping value}
}
\description{
Similar calculation to correlation but is not affected by high differences 
across values.
}
