% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{qc_protein_signal}
\alias{qc_protein_signal}
\title{Add QC flags to feature and protocol data simultaneously}
\usage{
qc_protein_signal(
  object,
  neg.names,
  targetAnnotations = NULL,
  neg.thresh = 20,
  qccols = c("#00008B80", "#FF000080")
)
}
\description{
Add QC flags to feature and protocol data simultaneously
}
