% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-autoplot.R
\name{qcProteinSignal}
\alias{qcProteinSignal}
\title{Generate Protein QC signal boxplot figure}
\usage{
qcProteinSignal(object, neg.names = NULL)
}
\arguments{
\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{neg.names}{names of IgGs, if NULL IgGs will be detected automatically}
}
\value{
protein names in the order of the figure
}
\description{
For use with protein data ONLY.
}
\examples{
testData <- readRDS(file= system.file("extdata","DSP_Proteogenomics_Example_Data", 
"proteinData.rds", package = "GeomxTools"))

proteinData <- analyteSubset(object = aggTestData, analyte = "protein")

igg.names <- iggNames(proteinData)

proteinOrder <- qcProteinSignal(object = proteinData, neg.names = igg.names)

}
