% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-autoplot.R
\name{snrOrder}
\alias{snrOrder}
\title{Generate ordered SNR matrix}
\usage{
snrOrder(object, neg.names)
}
\arguments{
\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{neg.names}{names of IgGs, if NULL IgGs will be detected automatically}
}
\value{
SNR matrix in increasing order
}
\description{
For use with protein data ONLY.
}
