% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{to.Seurat}
\alias{to.Seurat}
\title{Convert GeoMxSet Object to SeuratObject}
\usage{
to.Seurat(object, ident, normData = NULL, coordinates = NULL, forceRaw = FALSE)
}
\arguments{
\item{object}{GeoMxSet object to convert}

\item{ident}{column in GeoMxSet segmentProperties to set as Seurat object's identity class}

\item{normData}{assay containing normalized data}

\item{coordinates}{X and Y coordinates of each ROI, format: c(X,Y)}

\item{forceRaw}{should raw data be forced into SeuratObject, not recommended}
}
\value{
SeuratObject containing GeoMx data
}
\description{
Convert GeoMxSet Object to SeuratObject
}
\examples{
datadir <- system.file("extdata", "DSP_NGS_Example_Data", package = "GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))

target_demoData <- aggregateCounts(demoData)

target_demoData <- normalize(target_demoData, "quant")

seurat_demoData <- to.Seurat(object = target_demoData, ident = "cell_line", 
                             normData = "exprs_norm", forceRaw = FALSE)

}
