% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertors.R
\name{to.SpatialExperiment}
\alias{to.SpatialExperiment}
\title{Convert GeoMxSet Object to SpatialExperiment}
\usage{
to.SpatialExperiment(
  object,
  normData = "exprs_norm",
  coordinates = NULL,
  forceRaw = FALSE
)
}
\arguments{
\item{object}{GeoMxSet object to convert}

\item{normData}{assay containing normalized data}

\item{coordinates}{X and Y coordinates of each ROI, format: c(X,Y)}

\item{forceRaw}{should raw data be forced into SpatialExperiment, not recommended}
}
\value{
SpatialExperiment containing GeoMx data
}
\description{
Convert GeoMxSet Object to SpatialExperiment
}
\examples{
datadir <- system.file("extdata", "DSP_NGS_Example_Data", package = "GeomxTools")
demoData <- readRDS(file.path(datadir, "/demoData.rds"))

target_demoData <- aggregateCounts(demoData)

target_demoData <- normalize(target_demoData, "quant")

seurat_demoData <- to.SpatialExperiment(object = target_demoData, normData = "exprs_norm", 
                                        forceRaw = FALSE)

}
