% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{updateGeoMxSet}
\alias{updateGeoMxSet}
\title{Update GeoMxSet object to current version}
\usage{
updateGeoMxSet(object)
}
\arguments{
\item{object}{GeoMxSet object to update}
}
\value{
updated GeoMxSet object
}
\description{
Update GeoMxSet object to current version
}
