% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-normalize.R
\name{checkQCFlags,NanoStringGeoMxSet-method}
\alias{checkQCFlags,NanoStringGeoMxSet-method}
\title{checkQCFlags}
\usage{
\S4method{checkQCFlags}{NanoStringGeoMxSet}(object, removeLowLocalOutliers = FALSE, ...)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object to check the QC Flags}

\item{removeLowLocalOutliers}{logical, if TRUE it sets outlier counts to zero,  default is FALSE,}

\item{\code{ldots}}{optional arguments}
}
\value{
NanoStringGeoMxSet
}
\description{
checkQCFlags
}
\examples{
datadir <- system.file("extdata", "DSP_NGS_Example_Data",
  package = "GeomxTools"
)
demoData <- readRDS(file.path(datadir, "/demoData.rds"))
QCobject <- checkQCFlags(demoData)
}
