% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-autoplot.R
\name{concordancePlot}
\alias{concordancePlot}
\title{Produce concordance pairs plots}
\usage{
concordancePlot(
  mat,
  col = rgb(0, 0, 0, 0.5),
  collegend = NULL,
  legend.main = NULL,
  main = "",
  pch = 16,
  cex = 1.5,
  ...
)
}
\arguments{
\item{mat}{Matrix of values to be compared. Could be e.g. IgG probe counts,
or different normalization factors. Each variable/factor is in a column.}

\item{col}{Vector of colors for points}

\item{collegend}{Named vector of colors, used to draw a legend.}

\item{legend.main}{Title for the color legend, used to name the variable of interest}

\item{main}{Plot title}

\item{pch}{point type argument passed to pairs()}

\item{cex}{point size argument passed to pairs()}

\item{...}{Additional arguments passed to pairs()}
}
\description{
Draws a pairs plot of concordance between multiple factors/ probes
Upper panels are the concordance plot.
Lower panels are the standard deviation of the log2-ratios between the targets
}
\examples{
# simulate HKs:
x = pmax(rnorm(100, 50, 10), 0)
mat = sweep(matrix(rnorm(300, 0, 5), 100), 1, x, "+")
concordancePlot(mat = mat, col = rep(c("blue", "orange"), each = 50))
}
