% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Add QC flags to feature and protocol data simultaneously}
\usage{
plot_concordance(geneList, object, plot_factors)
}
\description{
Add QC flags to feature and protocol data simultaneously
}
