% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Generate concordance figure of genes based on user provided factors}
\usage{
plot_concordance(geneList, object, plot_factors)
}
\arguments{
\item{geneList}{names of genes to plot concordance, normally IgGs.}

\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{plot_factors}{segment factor to color the plot by}
}
\description{
For use with protein data ONLY.

Upper panels are the concordance plot.
Lower panels are the standard deviation of the log2-ratios between the genes.
}
\examples{
testData <- readRDS(file= system.file("extdata","DSP_Proteogenomics_Example_Data", 
"proteinData.rds", package = "GeomxTools"))

proteinData <- analyteSubset(object = aggTestData, analyte = "protein")

igg.names <- igg_names(proteinData)

qc_protein_signal(object = proteinData, neg.names = igg.names)

}
