% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{qc_protein_signal}
\alias{qc_protein_signal}
\title{Generate Protein QC signal boxplot figure}
\usage{
qc_protein_signal(object, neg.names = NULL)
}
\arguments{
\item{object}{name of the object class to subset
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{neg.names}{names of IgGs, if NULL IgGs will be detected automatically}
}
\description{
For use with protein data ONLY.
}
\examples{
testData <- readRDS(file= system.file("extdata","DSP_Proteogenomics_Example_Data", 
"proteinData.rds", package = "GeomxTools"))

proteinData <- analyteSubset(object = aggTestData, analyte = "protein")

igg.names <- igg_names(proteinData)

qc_protein_signal(object = proteinData, neg.names = igg.names)

}
