% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoStringGeoMxSet-qc.R
\name{setQCFlags,NanoStringGeoMxSet-method}
\alias{setQCFlags,NanoStringGeoMxSet-method}
\title{Add QC flags to feature or protocol data}
\usage{
\S4method{setQCFlags}{NanoStringGeoMxSet}(object, qcCutoffs = DEFAULTS, ...)
}
\arguments{
\item{object}{name of the object class to perform QC on
\enumerate{
    \item{NanoStringGeoMxSet, use the NanoStringGeoMxSet class}
}}

\item{qcCutoffs}{list of cutoffs and thresholds to use for QC}

\item{dataDim}{the dimension of the object to QC on
\enumerate{
    \item{sample, QC data on the AOI level}
    \item{feature, QC data on probe or target level}
}}
}
\value{
the object that QC was performed on
}
\description{
Add QC flags to feature or protocol data
}
\examples{

}
