% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shiftCountsOne}
\alias{shiftCountsOne}
\title{Add one to all counts in an expression matrix}
\usage{
shiftCountsOne(object, elt = "exprs")
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet object}

\item{elt}{expression matrix element in /code{assayDataElement}
to shift all counts by}
}
\value{
object of NanoStringGeoMxSet class
}
\description{
Add one to all counts in an expression matrix
}
\examples{
shiftCountsOne(demoData)

}
